tests: camada int/mart

+++ dags/dbt/airbnb/models/intermediate/int_airbnb_reviews.sql
@@ -3,9 +3,9 @@
 WITH int_airbnb_reviews AS (
     SELECT DISTINCT
         listing_id,
-        CAST(id AS BIGINT) AS review_id,
+        CAST(id AS INTEGER) AS review_id,
         CAST(date AS DATE) AS review_date,
-        CAST(reviewer_id AS BIGINT) AS reviewer_id,
+        CAST(reviewer_id AS INTEGER) AS reviewer_id,
         INITCAP(reviewer_name) AS reviewer_name,
         TRIM(REPLACE(REPLACE(comments, CHR(10), ' '), CHR(13), ' ')) AS comments
     FROM {{ ref('stg_airbnb_reviews') }}
